+++ tests/factories.py
@@ -2,10 +2,112 @@
 
 from __future__ import annotations
 
+import subprocess
+import wave
+from collections.abc import Callable
 from datetime import UTC, datetime
+from io import BytesIO
+from pathlib import Path
 from typing import Any
 
 from server.agent_runtime.models import SessionMeta
+
+
+def make_translator(locale: str = "zh") -> Callable[..., str]:
+    """Create a translator function bound to a fixed locale for testing."""
+    from lib.i18n import _ as i18n_translate
+
+    def translate(key: str, **kwargs) -> str:
+        return i18n_translate(key, locale=locale, **kwargs)
+
+    return translate
+
+
+def wav_bytes(duration_seconds: float, sample_rate: int = 8000) -> bytes:
+    """纯 stdlib 生成 wav 字节（不依赖 ffmpeg），供不要求真实音频编解码的用例使用。"""
+    buf = BytesIO()
+    with wave.open(buf, "wb") as wf:
+        wf.setnchannels(1)
+        wf.setsampwidth(2)
+        wf.setframerate(sample_rate)
+        wf.writeframes(b"\x00\x00" * int(duration_seconds * sample_rate))
+    return buf.getvalue()
+
+
+def make_test_video(path: Path, *, duration_sec: float = 1.0, fps: int = 30) -> None:
+    """使用 ffmpeg 生成极短测试视频（64x64 像素）"""
+    path.parent.mkdir(parents=True, exist_ok=True)
+    subprocess.run(
+        [
+            "ffmpeg",
+            "-y",
+            "-f",
+            "lavfi",
+            "-i",
+            f"color=black:size=64x64:duration={duration_sec}:rate={fps}",
+            "-c:v",
+            "libx264",
+            "-pix_fmt",
+            "yuv420p",
+            str(path),
+        ],
+        capture_output=True,
+        check=True,
+    )
+
+
+def make_test_video_with_audio_tail(
+    path: Path,
+    *,
+    video_duration_sec: float = 1.0,
+    audio_duration_sec: float = 1.5,
+    fps: int = 30,
+) -> None:
+    """生成音轨/容器尾部比视频轨更长的极短 MP4。"""
+    path.parent.mkdir(parents=True, exist_ok=True)
+    subprocess.run(
+        [
+            "ffmpeg",
+            "-y",
+            "-f",
+            "lavfi",
+            "-i",
+            f"color=black:size=64x64:duration={video_duration_sec}:rate={fps}",
+            "-f",
+            "lavfi",
+            "-i",
+            f"sine=frequency=440:duration={audio_duration_sec}",
+            "-c:v",
+            "libx264",
+            "-pix_fmt",
+            "yuv420p",
+            "-c:a",
+            "aac",
+            str(path),
+        ],
+        capture_output=True,
+        check=True,
+    )
+
+
+def make_test_audio(path: Path, *, duration_sec: float = 1.0) -> None:
+    """使用 ffmpeg 生成极短测试音频（正弦波 wav，pcm_s16le 为 ffmpeg 内置编码器）"""
+    path.parent.mkdir(parents=True, exist_ok=True)
+    subprocess.run(
+        [
+            "ffmpeg",
+            "-y",
+            "-f",
+            "lavfi",
+            "-i",
+            f"sine=frequency=440:duration={duration_sec}",
+            "-c:a",
+            "pcm_s16le",
+            str(path),
+        ],
+        capture_output=True,
+        check=True,
+    )
 
 
 def make_session_meta(**overrides) -> SessionMeta:
